requirements.txt
# BankSight: Transaction Intelligence Dashboard
# Requirements and Dependencies

# Core Dependencies
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0

# Data Analysis & Machine Learning
scikit-learn==1.3.1
scipy==1.11.2
statsmodels==0.14.0

# Visualization
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.12.2

# API & Web
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0

# Data Export
openpyxl==3.1.2
xlsxwriter==3.1.2
reportlab==4.0.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
validators==0.22.0

# Cryptography & Security
cryptography==41.0.3
pycryptodome==3.18.0

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
black==23.9.1
flake8==6.1.0
pylint==3.0.0
mypy==1.5.1
isort==5.12.0

# Production
gunicorn==21.2.0
uvicorn==0.23.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
